requirements.txt
numpy
pandas
scikit-learn
tensorflow
keras
joblib
pyTelegramBotAPI
